+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/mapping/model/MetaStoreMappingImpl.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2020 Expedia, Inc.
+ * Copyright (C) 2016-2021 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,6 +20,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
 
+import org.apache.hadoop.hive.metastore.MetaStoreFilterHook;
 import org.apache.hadoop.hive.metastore.api.AlreadyExistsException;
 import org.apache.hadoop.hive.metastore.api.Database;
 import org.apache.hadoop.hive.metastore.api.InvalidObjectException;
@@ -43,6 +44,7 @@ class MetaStoreMappingImpl implements MetaStoreMapping {
   private final AccessControlHandler accessControlHandler;
   private final String name;
   private final long latency;
+  private final MetaStoreFilterHook metastoreFilter;
 
   private final ConnectionType connectionType;
 
@@ -52,13 +54,15 @@ class MetaStoreMappingImpl implements MetaStoreMapping {
       CloseableThriftHiveMetastoreIface client,
       AccessControlHandler accessControlHandler,
       ConnectionType connectionType,
-      long latency) {
+      long latency,
+      MetaStoreFilterHook metastoreFilter) {
     this.databasePrefix = databasePrefix;
     this.name = name;
     this.client = client;
     this.accessControlHandler = accessControlHandler;
     this.connectionType = connectionType;
     this.latency = latency;
+    this.metastoreFilter = metastoreFilter;
   }
 
   @Override
@@ -74,6 +78,11 @@ class MetaStoreMappingImpl implements MetaStoreMapping {
   @Override
   public ThriftHiveMetastore.Iface getClient() {
     return client;
+  }
+
+  @Override
+  public MetaStoreFilterHook getMetastoreFilter() {
+    return metastoreFilter;
   }
 
   @Override