+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/server/WaggleDanceServerException.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2019 Expedia, Inc.
+ * Copyright (C) 2016-2021 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -25,6 +25,10 @@ public class WaggleDanceServerException extends RuntimeException {
 
   public WaggleDanceServerException(String message) {
     super(message);
+  }
+
+  public WaggleDanceServerException(String message, Throwable cause) {
+    super(message, cause);
   }
 
 }