+++ waggle-dance-core/src/test/java/com/hotels/bdp/waggledance/mapping/service/impl/YamlFederatedMetaStoreStorageTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2020 Expedia, Inc.
+ * Copyright (C) 2016-2021 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -160,6 +160,7 @@ public class YamlFederatedMetaStoreStorageTest {
     FederatedMetaStore metaStore = newFederatedInstance("hcom_2", "thrift://localhost:29083");
     metaStore.setDatabasePrefix("hcom_2_prefix_");
     assertThat(storage.getAll().get(2), is(metaStore));
+    assertThat(storage.getAll().get(2).getHiveMetastoreFilterHook(), is("filter.hook.class"));
   }
 
   @Test
@@ -209,10 +210,11 @@ public class YamlFederatedMetaStoreStorageTest {
     MappedTables mappedTables1 = new MappedTables("db1", Lists.newArrayList("tbl1"));
     MappedTables mappedTables2 = new MappedTables("db2", Lists.newArrayList("tbl2"));
     newFederatedInstance.setMappedTables(Lists.newArrayList(mappedTables1, mappedTables2));
+    newFederatedInstance.setHiveMetastoreFilterHook("filter.hook.class");
     storage.insert(newFederatedInstance);
     storage.saveFederation();
     List<String> lines = Files.readAllLines(file.toPath(), StandardCharsets.UTF_8);
-    assertThat(lines.size(), is(23));
+    assertThat(lines.size(), is(24));
     assertThat(lines.get(0), is("primary-meta-store:"));
     assertThat(lines.get(1), is("  access-control-type: READ_ONLY"));
     assertThat(lines.get(2), is("  database-prefix: ''"));
@@ -223,19 +225,20 @@ public class YamlFederatedMetaStoreStorageTest {
     assertThat(lines.get(7), is("- access-control-type: READ_ONLY"));
     assertThat(lines.get(8), is("  database-name-mapping: {}"));
     assertThat(lines.get(9), is("  database-prefix: hcom_2_"));
-    assertThat(lines.get(10), is("  latency: 0"));
-    assertThat(lines.get(11), is("  mapped-databases:"));
-    assertThat(lines.get(12), is("  - db1"));
-    assertThat(lines.get(13), is("  - db2"));
-    assertThat(lines.get(14), is("  mapped-tables:"));
-    assertThat(lines.get(15), is("  - database: db1"));
-    assertThat(lines.get(16), is("    mapped-tables:"));
-    assertThat(lines.get(17), is("    - tbl1"));
-    assertThat(lines.get(18), is("  - database: db2"));
-    assertThat(lines.get(19), is("    mapped-tables:"));
-    assertThat(lines.get(20), is("    - tbl2"));
-    assertThat(lines.get(21), is("  name: hcom_2"));
-    assertThat(lines.get(22), is("  remote-meta-store-uris: thrift://localhost:29083"));
+    assertThat(lines.get(10), is("  hive-metastore-filter-hook: filter.hook.class"));
+    assertThat(lines.get(11), is("  latency: 0"));
+    assertThat(lines.get(12), is("  mapped-databases:"));
+    assertThat(lines.get(13), is("  - db1"));
+    assertThat(lines.get(14), is("  - db2"));
+    assertThat(lines.get(15), is("  mapped-tables:"));
+    assertThat(lines.get(16), is("  - database: db1"));
+    assertThat(lines.get(17), is("    mapped-tables:"));
+    assertThat(lines.get(18), is("    - tbl1"));
+    assertThat(lines.get(19), is("  - database: db2"));
+    assertThat(lines.get(20), is("    mapped-tables:"));
+    assertThat(lines.get(21), is("    - tbl2"));
+    assertThat(lines.get(22), is("  name: hcom_2"));
+    assertThat(lines.get(23), is("  remote-meta-store-uris: thrift://localhost:29083"));
   }
 
   @Test