+++ waggle-dance-integration-tests/src/main/java/com/hotels/bdp/waggledance/WaggleDanceRunner.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2020 Expedia, Inc.
+ * Copyright (C) 2016-2021 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -216,6 +216,11 @@ public class WaggleDanceRunner implements WaggleDance.ContextListener {
 
     public Builder withPrimaryMappedTables(List<MappedTables> mappableTables) {
       primaryMetaStore.setMappedTables(mappableTables);
+      return this;
+    }
+
+    public Builder withHiveMetastoreFilterHook(String hiveMetastoreFilterHook) {
+      primaryMetaStore.setHiveMetastoreFilterHook(hiveMetastoreFilterHook);
       return this;
     }